Drafting processing.log that will contain processing info
This change introduces logic that will open file from disk read it's contents
and ensures that the players have persistent anonymization that is kept
on the processing machine. processing.log will also contain a list of files
that were already processed so that the program does not process the same
files twice in case of a crash.