Drafted profiler code, attempting parallelism fix
Currently after performing all of the tasks the program did not shut down.
This might be due to waitGroup not knowing when the task is done.
Added a piece of code wg.Done() that should alleviate this issue.

Drafted pprof profiler code that (if it works) could be used to optimize
the current solution.