Simplified map downloading pipeline
This change removed the persistence checks to verify if a replay was
previously processed as a part of the map downloading pipeline.
Currently all of the input replays are read to get the URL and map name,
further the downloader is seeded with excluding the maps that exist on
drive. All of the URLs that are left are downloaded.